Migrate the 'govet' hook to invoke
